+++ Changelog.md
@@ -1,5 +1,6 @@
 -----------Changelog for v1.0.8-----------
 
+- Improved mod compatibility by marking vanished players as spectators for other mods
 - Fixed fake join/leave messages not getting sent on discord servers via minecraft2discord
 - Fixed that this mod deleted persistent player data from other mods